Dyn prio should now work; color changes

+++ source/rp/rp_dyn_prio.h
@@ -6,14 +6,14 @@
 #define RP_DYN_PRIO_FRAME_COUNT 2
 struct rp_dyn_prio_t {
     struct rp_dyo_prio_screen_t {
-        u16 frame_size_acc;
+        u32 frame_size_acc;
         u16 priority_size_acc;
 
         u8 frame_size_n;
-        u8 frame_size[RP_DYN_PRIO_FRAME_COUNT];
+        u32 frame_size[RP_DYN_PRIO_FRAME_COUNT];
         u8 priority_size[RP_DYN_PRIO_FRAME_COUNT];
 
-        u8 frame_size_chn;
+        u32 frame_size_chn;
         u8 priority_size_chn;
 
         u8 frame_index;
@@ -23,7 +23,7 @@ struct rp_dyn_prio_t {
         u8 initializing;
         u8 priority;
 
-        u16 frame_size_est;
+        u32 frame_size_est;
         u16 priority_size_est;
     } s[SCREEN_COUNT];
     rp_lock_t mutex;
